исправил архитектуру в data слое
