+++ app/src/main/java/com/example/bankapp/data/repository/HistoryTransactionRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.example.bankapp.core.ResultState
 import com.example.bankapp.data.network.api.ApiService
 import com.example.bankapp.domain.model.TransactionDetailed
 import com.example.bankapp.domain.repository.HistoryTransactionRepository
-import com.example.bankapp.utils.safeApiCall
+import com.example.bankapp.data.utils.safeApiCall
 import javax.inject.Inject
 
 /**
@@ -31,17 +31,19 @@ class HistoryTransactionRepositoryImpl @Inject constructor(
      * @return [ResultState] с результатом запроса.
      */
     override suspend fun loadHistoryTransaction(
-        accountId: Int,
+        accountId: Int?,
         startDate: String,
         endDate: String
     ): ResultState<List<TransactionDetailed>> {
+
+
 
         return safeApiCall(
             mapper = {
                 it.toTransactionDetailed()
             },
             block = {apiService.getTransactions(
-                accountId = accountId,
+                accountId = accountId!!,
                 startDate = startDate,
                 endDate = endDate
             )
