+++ app/src/main/java/com/example/bankapp/data/repository/TodayTransactionRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.example.bankapp.core.ResultState
 import com.example.bankapp.data.network.api.ApiService
 import com.example.bankapp.domain.model.Transaction
 import com.example.bankapp.domain.repository.TodayTransactionRepository
-import com.example.bankapp.utils.safeApiCall
+import com.example.bankapp.data.utils.safeApiCall
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 import javax.inject.Inject
@@ -31,6 +31,9 @@ class TodayTransactionRepositoryImpl @Inject constructor(
      * @return [ResultState] с результатом запроса.
      */
     override suspend fun loadTodayTransaction(accountId: Int?): ResultState<List<Transaction>> {
+
+
+
         return safeApiCall(
             mapper = {
                 val transaction = it.toTransaction()
